pyproject.toml
[project]
name = "voicekey"
version = "0.1.0"
description = "Voice dictation for macOS. Hold Option, speak, release. Text appears at your cursor."
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"
authors = [
    { name = "Adam Khakhar", email = "adam.khakhar@gmail.com" },
]
keywords = [
    "voice-dictation",
    "transcription",
    "macos",
    "openai",
    "speech-to-text",
    "cli",
    "dictation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1",
    "keyring>=25.0",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "httpx>=0.27",
    "rumps>=0.4",
    "tomli-w>=1.0",
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
    "rich>=13.0",
]

[project.scripts]
voicekey = "voicekey.cli:main"

[project.urls]
Homepage = "https://github.com/adamkhakhar/voicekey"
Repository = "https://github.com/adamkhakhar/voicekey"
Issues = "https://github.com/adamkhakhar/voicekey/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
